packages/ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
boto3==1.34.25
numpy==1.26.3
scipy==1.12.0
Pillow==10.2.0
rq==1.16.1
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.6
